+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     println("At least Mimic 0.6 is required. Please download it from {link here}")
   }
   ```
+- Add optional payload to `ItemsRegistry.getItem`. It may be used to customize item.
 
 ### Bukkit Plugin